+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.2 (unreleased)
 
+- Added `verbose` option
 - Fixed issue with stopping too early
 
 ## 0.1.1 (2021-10-17)